Move test dot to Digraph
